Move PoweredByAdapter to Overview package
